requirements.txt
flask==2.0.1
werkzeug==2.0.3
fastapi>=0.100.0
uvicorn>=0.15.0
tensorflow==2.18.0
psycopg2-binary==2.9.10
langchain==0.3.19
pgvector==0.1.8
openai==1.3.0
sqlalchemy==1.4.40
python-dotenv>=0.21.0
numpy>=1.26.0
# faiss-cpu==1.7.2  # Comment out or replace with: hnswlib
requests==2.28.1
ipaddress==1.0.23
python-multipart==0.0.5
pydantic>=2.7.4,<3.0.0
# tiktoken==0.3.3   # Try installing separately with --only-binary